add env selector for the visualizer
